add loader
